src/test/javascript/spec/app/entities/fincasMs/herramienta/herramienta.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HerramientaService } from 'app/entities/fincasMs/herramienta/herramienta.service';
import { IHerramienta, Herramienta } from 'app/shared/model/fincasMs/herramienta.model';
import { TipoHerramienta } from 'app/shared/model/enumerations/tipo-herramienta.model';

describe('Service Tests', () => {
  describe('Herramienta Service', () => {
    let injector: TestBed;
    let service: HerramientaService;
    let httpMock: HttpTestingController;
    let elemDefault: IHerramienta;
    let expectedResult: IHerramienta | IHerramienta[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(HerramientaService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Herramienta(0, 'AAAAAAA', TipoHerramienta.VEHICULO, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Herramienta', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Herramienta()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Herramienta', () => {
        const returnedFromService = Object.assign(
          {
            nombre: 'BBBBBB',
            tipo: 'BBBBBB',
            descripcion: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Herramienta', () => {
        const returnedFromService = Object.assign(
          {
            nombre: 'BBBBBB',
            tipo: 'BBBBBB',
            descripcion: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Herramienta', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
